Split main - initial approach
